[#11833] refactor(core): Move PagedResult from api to core authorization
Keep paginated list results as an internal core type rather than a
public api module type, per review feedback.
